CHORE: change directory structure and file name in naver
The previous structure assumed that the view service and influencer service would have
diffrent interfaces, but as progressed, i realised that it would be more efficent
for the view service and influencer service to have the same interface.
Therefore, i renamed the existing ViewService interface to NaverService.

Also, I changed the structure to put all the files under the naver directory
because i don't think i need a lot of detailed files in the naver/view and naver/influencer directory