feat: modify top 10 list filtering method in naver view service
Calculating based on recommendedPostCount is fragile and error-prune
since the number of recommended post is not deterministic.

For example, assume there are 5 recommended posts among the top 10 <li>
elements. In this case, the search range is expanded to 15 to check if
any of the posts contain a maching URL.

However, if there are additional recommended posts among the 10-15
range, they are excluded from the count. As a result, using nth-child
prevents checking exactly 10 posts.

For example, if there are 2 additional recommended posts in the 10-15
range, only the top 8 posts will be examined instead of 10.

To address this, I modified the logic for selecting the top 10 posts.
Instaed of relying on nth-child, I first retrieve all non-recommended
posts and then use slice to extract the top 10 before performing
examination.