feat: returns random first div as fallback when there are no search results in instagram search page
as is: throws an error when the search results is empty in instagram
search which results in crash the application.

to be: returns first found div as a fallback. returns as Array<div> form
to be capable with previous interface, as it is used by findPost method
and screenshot method. Therefore, there is no need to modify the
findPost and screenshot method.